+++ hutool-core/src/main/java/cn/hutool/core/convert/impl/BooleanConverter.java
@@ -1,6 +1,7 @@
 package cn.hutool.core.convert.impl;
 
 import cn.hutool.core.convert.AbstractConverter;
+import cn.hutool.core.util.BooleanUtil;
 
 /**
  * 波尔转换器
@@ -15,7 +16,7 @@ public class BooleanConverter extends AbstractConverter<Boolean>{
 			return Boolean.valueOf((boolean)value);
 		}
 		String valueStr = convertToStr(value);
-		return Boolean.valueOf(PrimitiveConverter.parseBoolean(valueStr));
+		return Boolean.valueOf(BooleanUtil.toBoolean(valueStr));
 	}
 
 }